New Entities Created

+++ back-end/FSM/src/main/java/com/FantasyStockMarket/FSM/Entity/User/User.java
@@ -18,7 +18,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    int id;
+    Long id;
 
     @Column(name = "email_id")
     String emailId;
@@ -31,11 +31,12 @@ public class User {
     @Column(name = "joined_date", insertable = false)
     Timestamp joinedDate;
 
-    public int getId() {
+
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
